chore: factor out util, rpc functions from _workers.py
